Add Flea Market

+++ HQCsOneBlock/src/main/java/org/hqcplays/hqcsoneblock/PlayerSaveData.java
@@ -10,7 +10,7 @@ import java.util.Map;
 import java.util.Set;
 
 public class PlayerSaveData implements Serializable {
-    public int balance;
+    public double balance;
     public Map<Material, Double> blockChances;
     public Set<Material> unlockedBlocks;
     public ArrayList<FleaListing> fleaListings;
